requirements.txt
click>=8.0
requests>=2.28
beautifulsoup4>=4.11
python-dotenv>=1.0
rich>=13.0
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
nltk>=3.8
selenium>=4.0
webdriver-manager>=4.0
playwright>=1.40
